remove old e-ann, starting to combine submits into a single POST request
