+++ tinker/__init__.py
@@ -39,7 +39,6 @@ from tinker.events.views import event_blueprint
 from tinker.faculty_bio.views import faculty_bio_blueprint
 from tinker.admin.redirects.views import redirect_blueprint
 from tinker.heading_upgrade.views import heading_upgrade
-from tinker.e_announcements.views import e_announcements_blueprint
 from tinker.admin.sync.views import sync_blueprint
 from tinker.admin.publish.views import publish_blueprint
 from tinker.admin.roles.views import blink_roles_blueprint
@@ -48,17 +47,17 @@ from tinker.admin.cache.views import cache_blueprint
 app.register_blueprint(event_blueprint, url_prefix='/event')
 app.register_blueprint(faculty_bio_blueprint, url_prefix='/faculty-bio')
 app.register_blueprint(heading_upgrade, url_prefix='/heading-upgrade')
-app.register_blueprint(e_announcements_blueprint, url_prefix='/e-announcement')
 app.register_blueprint(sync_blueprint, url_prefix='/admin/sync')
 app.register_blueprint(publish_blueprint, url_prefix='/admin/publish-manager')
 app.register_blueprint(blink_roles_blueprint, url_prefix='/admin/blink-roles')
 app.register_blueprint(cache_blueprint, url_prefix='/admin/cache-clear')
 app.register_blueprint(redirect_blueprint, url_prefix='/admin/redirect')
 
+
 # New importing of routes and blueprints
-from tinker.e_announcements_new import *
-NewEAnnouncementsView.register(NewEAnnouncementsBlueprint)
-app.register_blueprint(NewEAnnouncementsBlueprint, url_prefix='/test')
+from tinker.e_announcements import *
+EAnnouncementsView.register(EAnnouncementsBlueprint)
+app.register_blueprint(EAnnouncementsBlueprint)
 
 CsrfProtect(app).exempt(redirect_blueprint)
