--- tinker/e_announcements/templates/e-announcements-confirm-edit.html
@@ -1,20 +1,0 @@
-{% extends "tinker-base.html" %}
-
-{% block title %}Congrats!{% endblock %}
-
-
-
-{% block main_content %}
-
-  <div class="row">
-    <div class="large-12 columns">
-        <h1>Congrats!</h1>
-        <p>
-          You've successfully edited your E-Announcement. Once your E-Announcement has been approved, it will appear on your Tinker <a href="/e-announcement">E-Announcement page</a> where you'll be able to view the announcement page or make changes to it.
-        </p>
-         <p>If you have any questions, please contact the Office of Communications and Marketing at <a href="mailto:e-announcements@bethel.edu">e-announcements@bethel.edu</a>.</p>
-
-    </div>
-  </div>
-
-{% endblock %}
