+++ tinker/e_announcements/templates/e-announcements-confirm.html
