--- tinker/e_announcements_new/banner_roles_mapping.py
@@ -1,19 +1,0 @@
-def get_banner_roles_mapping():
-    return [
-        'CAS',
-        'CAPS',
-        'GS',
-        'BSSP-TRADITIONAL',
-        'BSSP-DISTANCE',
-        'BSSD-TRADITIONAL',
-        'BSSD-DISTANCE',
-        'BSOE-TRADITIONAL',
-        'BSOE-DISTANCE',
-        'CAS',
-        'CAPS',
-        'GS',
-        'BSSP',
-        'BSSD',
-        'St. Paul',
-        'San Diego'
-    ]
