--- tinker/e_announcements_new/templates/e-announcements-delete-confirm.html
@@ -1,19 +1,0 @@
-{% extends "tinker-base.html" %}
-
-{% block title %}Success{% endblock %}
-
-{% block page_title %}Bethel University Tinker{% endblock %}
-
-{% block main_content %}
-
-  <div class="row">
-    <div class="large-12 columns">
-
-        <p>
-          Your E-Announcements has been deleted. It will be removed from your <a href="/e-announcement/">Tinker homepage</a> in a few minutes.
-        </p>
-
-    </div>
-  </div>
-
-{% endblock %}
