--- tinker/e_announcements_new/templates/e-announcements-home2.html
@@ -1,90 +1,0 @@
-{% extends "tinker-base.html" %}
-
-{% block title %}Home{% endblock %}
-
-{% block styles %}
-  <style type="text/css">
-  </style>
-{% endblock %}
-
-{% block page_title %}Bethel University Tinker{% endblock %}
-
-{% block main_content %}
-<div class="row">
-  <div class="large-12 columns">
-
-    <p>Below is the list of E-Announcements you have access to edit. These are sorted by the first date entered, newest dates first.
-    </p>
-  </div>
- </div>
-      <a href="{{ url_for('new-e-announcement.NewEAnnouncementsView:new') }}" class="small button success">Submit New E-Announcement</a>
-<hr/>
-
-<div class="row">
-
-  <div class="large-12 columns">
-
-   {% for form in forms%}
-      <div class="row">
-        <div class="large-6 columns">
-          <p>
-            {% if form.path != '' %}
-              <a href={{ form.path }}>{{ form.title }}</a>
-            {% else %}
-              {{ form.title }}
-            {% endif %}
-            <br/>
-            <small>{{ form.first_date | safe}}</small>
-            <br/>
-            <small>{{ form.second_date | safe}}</small>
-          </p>
-        </div>
-        <div class="large-6 columns small-3">
-          <ul class="button-group">
-            {% if form.workflow_status %}
-              <li>This announcement is pending approval.</li>
-            {% else %}
-              {% if not form.first_date_past and not form.second_date_past %}
-                <li><a href="{{ url_for('new-e-announcement.NewEAnnouncementsView:edit', e_announcement_id=form.id) }}" class="small button">Edit</a></li>
-              {% else %}
-                  <li><a href="/e-announcement/view/{{ form.id }}" class="small button">View</a></li>
-{#                  <li><a href="/e-announcement/duplicate/{{ form.id }}" class="small button">Duplicate</a></li>#}
-              {% endif %}
-              {% if not form.first_date_past %}
-                  <li><a href="#" data-reveal-id="confirmModal" id="{{ form.id }}" data-e-announcement-id="{{ form.id }}" class="small button alert delete-press">Delete</a></li>
-              {% endif %}
-            {% endif %}
-           </ul>
-        </div>
-        <hr/>
-      </div>
-    {% endfor %}
-
-
-  </div>
-</div>
-
-
-{#Delete confirm modal#}
-<div id="confirmModal" class="reveal-modal" data-reveal>
-  <h2>Are you sure you want to delete?</h2>
-  <p class="lead">Your E-Announcement will removed.</p>
-  <a class="modal-confirm" id="modal-confirm" href="/e-announcement/delete/formid">Delete</a>
-  <a class="close-reveal-modal">&#215;</a>
-</div>
-
-{% endblock %}
-
-{% block scripts %}
-
-{#  Script to update the modal submit to delete the correct form#}
-  <script type="text/javascript">
-    $(".delete-press").click(function(){
-      $("#modal-confirm").attr("href", "/e-announcement/delete/" + $(this).data('e-announcement-id'));
-    });
-  $(document).foundation();
-  </script>
-
-{% endblock %}
-
-
