--- tinker/e_announcements_new/templates/e-announcements-in-workflow.html
@@ -1,22 +1,0 @@
-{% extends "tinker-base.html" %}
-
-{% block title %}Pending Approval{% endblock %}
-
-{% block page_title %}Bethel University Tinker{% endblock %}
-
-{% block main_content %}
-
-  <div class="row">
-    <div class="large-12 columns">
-
-        <h3>
-          Edits Pending Approval
-        </h3>
-
-        <p>You recently made an edit to your E-Announcement and is currently pending approval. Please wait until the changes have been approved before you make additional edits. Go back to your <a href="/e-announcement">E-Announcements</a>.</p>
-        <p>If you have any questions or concerns about the submission of the E-Announcement, please contact Christine Ernston, at c-erntson@bethel.edu</p>
-
-    </div>
-  </div>
-
-{% endblock %}
