--- tinker/e_announcements_new/templates/e-announcements-view.html
@@ -1,21 +1,0 @@
-{% extends "tinker-base.html" %}
-
-{% block title %}{{ edit_data.title }}{% endblock %}
-{% block page_title %}{{ edit_data.title }}{% endblock %}
-
-{% block main_content %}
-
-    <p>{{ edit_data.message | safe}}</p>
-
-    <hr/>
-
-    <h5>First Date</h5>
-    <p>{{ first }}</p>
-    {% if second %}
-        <h5>Second Date</h5>
-        <p>{{ second }}</p>
-    {% endif %}
-    <small>{{ edit_data.banner_roles }}</small>
-
-{% endblock %}
-
